Remove implicit dependency on the facets gem
The `snakecase` method is not part of Rails or ActiveRecord but is
included in the "facets" gem, which however is not explicitely required
by nandi (though it's used indirectly by many other projects at GC).

This commit removes that implicit dependency by just porting the same
implementation of the `snakecase` method, allowing projects that do not
include that dependency to also use nandi.