Standardise on access_exclusive_X naming for config
While everyday users of Nandi should not have to care about what locks
their migration takes, with us guaranteeing that they won't take a table
out for an unreasonable time, we should not hide from people making
configuration changes the heuristics that we use to decide what a
reasonable timeout is. So this standardises the naming more explicitly.